fix#4355

+++ druid-admin/src/main/java/com/alibaba/druid/admin/model/dto/ConnectionResult.java
@@ -25,25 +25,25 @@ public class ConnectionResult {
     public static class ContentBean {
 
         @JSONField(name = "id")
-        private int id;
+        private long id;
         @JSONField(name = "connectionId")
-        private int connectionId;
+        private long connectionId;
         @JSONField(name = "useCount")
-        private int useCount;
+        private long useCount;
         @JSONField(name = "lastActiveTime")
         private String lastActiveTime;
         @JSONField(name = "connectTime")
         private String connectTime;
         @JSONField(name = "holdability")
-        private int holdability;
+        private long holdability;
         @JSONField(name = "transactionIsolation")
-        private int transactionIsolation;
+        private long transactionIsolation;
         @JSONField(name = "autoCommit")
         private boolean autoCommit;
         @JSONField(name = "readoOnly")
         private boolean readoOnly;
         @JSONField(name = "keepAliveCheckCount")
-        private int keepAliveCheckCount;
+        private long keepAliveCheckCount;
         @JSONField(name = "pscache")
         private List<?> pscache;
     }
